Add concern for Api::V1::Users::RecipesController
