+++ app/controllers/api/v1/users/recipes_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class Api::V1::Users::RecipesController < Api::V1::ApplicationController
+  include RecipeConcerns
   before_action :authenticate_user!
   before_action :ingredient_list, only: %i[create]
   after_action :tag_list, only: %i[create], if: -> { params[:tags].present? }
@@ -62,27 +63,27 @@ class Api::V1::Users::RecipesController < Api::V1::ApplicationController
     params.require(:recipe).permit(:name)[:name]
   end
 
-  def tag_list
-    @recipe.tags << Tag.create_list(params[:tags])
-  end
+  #def tag_list
+  #  @recipe.tags << Tag.create_list(params[:tags])
+  #end
 
-  def assign_family
-    @recipe.assign_family
-  end
+  #def assign_family
+  #  @recipe.assign_family
+  #end
 
-  def recipe_ingredient_list
-    @recipe.recipe_ingredient_list(recipe_ing_params)
-  end
+  #def recipe_ingredient_list
+  #  @recipe.recipe_ingredient_list(recipe_ing_params)
+  #end
 
-  def ingredient_list
-    Ingredient.create_list(ingredient_params)
-  end
+  #def ingredient_list
+  #  Ingredient.create_list(ingredient_params)
+  #end
 
-  def recipe_activity
-    user = current_user.id
-    client = Stream::Client.new(ENV['STREAM_KEY'], ENV['STREAM_SECRET'])
-    feed = client.feed('user', user)
-    activity_data = { actor: user, verb: 'post', object: 1, post: "#{current_user.name} created a new recipe" }
-    feed.add_activity(activity_data)
-  end
+  #def recipe_activity
+  #  user = current_user.id
+  #  client = Stream::Client.new(ENV['STREAM_KEY'], ENV['STREAM_SECRET'])
+  #  feed = client.feed('user', user)
+  #  activity_data = { actor: user, verb: 'post', object: 1, post: "#{current_user.name} created a new recipe" }
+  #  feed.add_activity(activity_data)
+  #end
 end
